% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BankAccount-exports.R
\name{BankAccount}
\alias{BankAccount}
\title{Create a BankAccount Object from the BankAccount C++ Class}
\arguments{
\item{starting_balance}{Amount in the BankAccount. Default: Start with 250.}
}
\value{
A \code{BankAccount} object from the \emph{C++} BankAccount Class.
}
\description{
Allows for the creation of a BankAccount Object in \emph{C++} from \emph{R}
using the \emph{C++} BankAccount class.
}
\examples{
##################
## Constructor

# Construct new BankAccount object called "jjb_bank"
jjb_bank = new(BankAccount)

##################
## Setters

jjb_bank$deposit(10)
jjb_bank$withdraw(20)

##################
## Getters

jjb_bank$get_current_balance

}
