% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppBankAccount-package.R
\docType{package}
\name{RcppBankAccount-package}
\alias{RcppBankAccount}
\alias{RcppBankAccount-package}
\title{RcppBankAccount: Expose the BankAccount C++ Class using Rcpp Modules}
\description{
Provides an implementation of a BankAccount 'C++' class being connected to 'R' using 'Rcpp Modules' to expose the underlying class methods.
}
\details{
RcppBankAccount is a package showing how to use Rcpp Modules to make available
a C++ class under S4.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rd.thecoatlessprofessor.com/rcpp-modules-bank-account/}
  \item Report bugs at \url{https://github.com/coatless-r-n-d/rcpp-modules-bank-account}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
